+++ app/test/_helpers/sync-engine.helper.ts
@@ -1,3 +1,5 @@
+import {jest} from '@jest/globals';
+import {AxiosResponse} from 'axios';
 import {iCloud} from "../../src/lib/icloud/icloud";
 import {iCloudPhotos} from "../../src/lib/icloud/icloud-photos/icloud-photos";
 import {PhotosLibrary} from "../../src/lib/photos-library/photos-library";
@@ -12,7 +14,7 @@ export function syncEngineFactory(): SyncEngine {
         },
         new iCloud({
             "username": Config.username,
-            "password": Config.password, 
+            "password": Config.password,
             "trustToken": Config.trustToken,
             "dataDir": Config.appDataDir,
         }),
@@ -21,5 +23,13 @@ export function syncEngineFactory(): SyncEngine {
         }),
     );
     syncEngine.iCloud.photos = new iCloudPhotos(syncEngine.iCloud.auth);
+    return syncEngine;
+}
+
+export function mockSyncEngineForAssetQueue(syncEngine: SyncEngine): SyncEngine {
+    syncEngine.photosLibrary.verifyAsset = jest.fn(() => false);
+    syncEngine.photosLibrary.writeAsset = jest.fn(async () => {});
+    syncEngine.photosLibrary.deleteAsset = jest.fn(async () => {});
+    syncEngine.iCloud.photos.downloadAsset = jest.fn(async () => ({} as AxiosResponse<any, any>));
     return syncEngine;
 }